chore: #1668 split test common utils
