+++ tee-worker/ts-tests/common/utils/storage.ts
@@ -1,0 +1,153 @@
+import { Metadata } from '@polkadot/types';
+import { HexString } from '@polkadot/util/types';
+import { u8aToHex, u8aConcat } from '@polkadot/util';
+import { xxhashAsU8a } from '@polkadot/util-crypto';
+import { StorageEntryMetadataV14, SiLookupTypeId, StorageHasherV14 } from '@polkadot/types/interfaces';
+import { sendRequest } from '../call';
+import { blake2128Concat, twox64Concat, identity } from '../helpers';
+import { IntegrationTestContext, IdentityContext } from '../type-definitions';
+
+const base58 = require('micro-base58');
+
+//sidechain storage utils
+export function buildStorageEntry(metadata: Metadata, prefix: string, method: string): StorageEntryMetadataV14 | null {
+    for (const pallet of metadata.asV14.pallets) {
+        if (pallet.name.toString() == prefix) {
+            const storage = pallet.storage.unwrap();
+
+            for (const item of storage.items) {
+                if (item.name.toString() == method) {
+                    return item;
+                }
+            }
+        }
+    }
+    return null;
+}
+
+export function buildStorageKey(
+    metadata: Metadata,
+    prefix: string,
+    method: string,
+    keyTypeId?: SiLookupTypeId,
+    hashers?: Array<StorageHasherV14>,
+    input?: Array<unknown>
+): Uint8Array {
+    let storageKey = u8aConcat(xxhashAsU8a(prefix, 128), xxhashAsU8a(method, 128));
+    if (keyTypeId && hashers && input) {
+        let keyTypeIds = hashers.length === 1 ? [keyTypeId] : metadata.registry.lookup.getSiType(keyTypeId).def.asTuple;
+        for (let i = 0; i < keyTypeIds.length; i++) {
+            const theKeyTypeId = keyTypeIds[i];
+            const theHasher = hashers[i].toString();
+            const theKeyItem = input[i];
+            // get the scale encoded input data by encoding the input
+            const theKeyType = metadata.registry.createLookupType(theKeyTypeId);
+            const theKeyItemEncoded = metadata.registry.createType(theKeyType, theKeyItem).toU8a();
+            // apply hasher
+            let theKeyItemAppliedHasher;
+            if (theHasher == 'Blake2_128Concat') {
+                theKeyItemAppliedHasher = blake2128Concat(theKeyItemEncoded);
+            } else if (theHasher == 'Twox64Concat') {
+                theKeyItemAppliedHasher = twox64Concat(theKeyItemEncoded);
+            } else if (theHasher == 'Identity') {
+                theKeyItemAppliedHasher = identity(theKeyItemEncoded);
+            } else {
+                throw new Error(`The hasher ${theHasher} is not support.`);
+            }
+            storageKey = u8aConcat(storageKey, theKeyItemAppliedHasher);
+        }
+    }
+    return storageKey;
+}
+export async function buildStorageHelper(
+    metadata: Metadata,
+    prefix: string,
+    method: string,
+    ...input: Array<unknown>
+): Promise<string | null> {
+    const storageEntry = buildStorageEntry(metadata, prefix, method);
+    if (!storageEntry) {
+        throw new Error('Can not find the storage entry from metadata');
+    }
+    let storageKey, valueType;
+
+    if (storageEntry.type.isPlain) {
+        storageKey = buildStorageKey(metadata, prefix, method);
+        valueType = metadata.registry.createLookupType(storageEntry.type.asPlain);
+    } else if (storageEntry.type.isMap) {
+        const { hashers, key, value } = storageEntry.type.asMap;
+        if (input.length != hashers.length) {
+            throw new Error('The `input` param is not correct');
+        }
+        storageKey = buildStorageKey(metadata, prefix, method, key, hashers, input);
+        valueType = metadata.registry.createLookupType(value);
+    } else {
+        throw new Error('Only support plain and map type');
+    }
+    console.debug(`storage key: ${u8aToHex(storageKey)}`);
+    return u8aToHex(storageKey);
+}
+
+export async function checkUserShieldingKeys(
+    context: IntegrationTestContext,
+    pallet: string,
+    method: string,
+    address: HexString
+): Promise<string> {
+    const storageKey = await buildStorageHelper(context.metaData, pallet, method, address);
+
+    let base58mrEnclave = base58.encode(Buffer.from(context.mrEnclave.slice(2), 'hex'));
+
+    let request = {
+        jsonrpc: '2.0',
+        method: 'state_getStorage',
+        params: [base58mrEnclave, storageKey],
+        id: 1,
+    };
+    let resp = await sendRequest(context.tee, request, context.api);
+
+    return resp.value;
+}
+
+export async function checkUserChallengeCode(
+    context: IntegrationTestContext,
+    pallet: string,
+    method: string,
+    address: HexString,
+    identity: HexString
+): Promise<string> {
+    const storageKey = await buildStorageHelper(context.metaData, pallet, method, address, identity);
+
+    let base58mrEnclave = base58.encode(Buffer.from(context.mrEnclave.slice(2), 'hex'));
+
+    let request = {
+        jsonrpc: '2.0',
+        method: 'state_getStorage',
+        params: [base58mrEnclave, storageKey],
+        id: 1,
+    };
+    let resp = await sendRequest(context.tee, request, context.api);
+    return resp.value;
+}
+
+export async function checkIDGraph(
+    context: IntegrationTestContext,
+    pallet: string,
+    method: string,
+    address: HexString,
+    identity: HexString
+): Promise<IdentityContext> {
+    const storageKey = await buildStorageHelper(context.metaData, pallet, method, address, identity);
+
+    let base58mrEnclave = base58.encode(Buffer.from(context.mrEnclave.slice(2), 'hex'));
+
+    let request = {
+        jsonrpc: '2.0',
+        method: 'state_getStorage',
+        params: [base58mrEnclave, storageKey],
+        id: 1,
+    };
+    let resp = await sendRequest(context.tee, request, context.api);
+    const IDGraph = context.api.createType('IdentityContext', resp.value).toJSON() as IdentityContext;
+    return IDGraph;
+}
